responses_copy/chat_data/requirements.txt
python-dotenv==1.0.0
openai==1.67.0
Flask==2.2.5
Werkzeug==2.2.3